requirements.txt
datasets
einops
hydra-core
matplotlib
numpy
pandas
pytorch_lightning
torch
torchaudio
torchvision
wandb
